+++ warehouse/metrics_tools/transformer/__init__.py
@@ -9,3 +9,5 @@ _before_ being stored in the sqlmesh state.
 
 from .base import *
 from .intermediate import *
+from .transformer import *
+from .qualify import *
